+++ braidpool-common/src/lib.rs
@@ -1,0 +1,3 @@
+//! Common utility module to be shared across crates
+pub mod cpunet;
+pub(crate) mod error;
